rgw: move RGWObjManifest logic out of rgw_dencoder.cc
now that RGWObjManifest has its own source file rgw_obj_manifest.cc,
rgw_dencoder.cc should only contain the logic related to encoding
